chore(ci): stub automated (prod) releases
this commit implements an initial workflow for prod releases for the cli.
it is capable of:
- allowing users to select a version and tag to publish
- generating a build based using the same logic as other ci workflows
- generating a prod build version

it does nothing of consequence, other than log information to the
console. once this has landed, it can be test manually to root out any
bugs in the implementation as it stands and complete the implementation